# server/doc/api/swagger.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/upload": {
      "post": {
        "description": "multipart/form-data 上传单文件，返回可访问 URL",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "bucket": {
                    "description": "public/private",
                    "type": "string"
                  },
                  "file": {
                    "description": "文件本体",
                    "format": "binary",
                    "type": "string"
                  },
                  "scene": {
                    "description": "业务场景标记",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "properties": {
                        "fileId": {
                          "description": "文件元数据 ID",
                          "type": "number"
                        },
                        "url": {
                          "description": "静态访问 URL",
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "单文件上传",
        "tags": [
          "upload"
        ]
      }
    },
    "/api/v1/upload/{fileId}": {
      "delete": {
        "description": "软删除文件元数据（仅本人或管理员）",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "删除文件（软删）",
        "tags": [
          "upload"
        ]
      },
      "get": {
        "description": "按 fileId 查询上传文件元数据",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "查询文件元数据",
        "tags": [
          "upload"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}